+++ eras/allegra/impl/testlib/Test/Cardano/Ledger/Allegra/Imp.hs
@@ -1,30 +1,23 @@
-{-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Test.Cardano.Ledger.Allegra.Imp (spec) where
 
-import Cardano.Ledger.Allegra (AllegraEra)
 import Cardano.Ledger.Core
 import Cardano.Ledger.Shelley.Rules
 import Test.Cardano.Ledger.Allegra.ImpTest
 import Test.Cardano.Ledger.Imp.Common
-import qualified Test.Cardano.Ledger.Shelley.Imp as ShelleyImp
+import qualified Test.Cardano.Ledger.Shelley.Imp as Shelley
 
 spec ::
-  forall era.
   ( ShelleyEraImp era
-  , EraSpecificSpec era
   , Event (EraRule "RUPD" era) ~ RupdEvent
   ) =>
+  proxy era ->
   Spec
-spec = do
-  ShelleyImp.spec @era
-
-instance EraSpecificSpec AllegraEra where
-  eraSpecificSpec = ShelleyImp.shelleyEraSpecificSpec
+spec era = do
+  Shelley.spec era
+  describe "AllegraEra Onwards" $ pure ()