+++ eras/alonzo/impl/testlib/Test/Cardano/Ledger/Alonzo/Imp.hs
@@ -1,51 +1,39 @@
-{-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE RankNTypes #-}
-{-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeOperators #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Test.Cardano.Ledger.Alonzo.Imp where
 
-import Cardano.Ledger.Alonzo (AlonzoEra)
 import Cardano.Ledger.Core
 import Cardano.Ledger.Shelley.Core (ShelleyEraTxCert)
 import Cardano.Ledger.Shelley.Rules
-import qualified Test.Cardano.Ledger.Alonzo.Imp.BbodySpec as Bbody
-import qualified Test.Cardano.Ledger.Alonzo.Imp.UtxoSpec as Utxo
-import qualified Test.Cardano.Ledger.Alonzo.Imp.UtxosSpec as Utxos
-import qualified Test.Cardano.Ledger.Alonzo.Imp.UtxowSpec as Utxow
+import qualified Test.Cardano.Ledger.Alonzo.Imp.BbodySpec as BBODY
+import qualified Test.Cardano.Ledger.Alonzo.Imp.UtxoSpec as UTXO
+import qualified Test.Cardano.Ledger.Alonzo.Imp.UtxosSpec as UTXOS
+import qualified Test.Cardano.Ledger.Alonzo.Imp.UtxowSpec as UTXOW
 import Test.Cardano.Ledger.Alonzo.ImpTest
 import Test.Cardano.Ledger.Imp.Common
-import qualified Test.Cardano.Ledger.Mary.Imp as MaryImp
-import qualified Test.Cardano.Ledger.Shelley.Imp as ShelleyImp
+import qualified Test.Cardano.Ledger.Mary.Imp as Mary
 
 spec ::
-  forall era.
   ( AlonzoEraImp era
-  , EraSpecificSpec era
   , Event (EraRule "RUPD" era) ~ RupdEvent
   ) =>
+  proxy era ->
   Spec
-spec = do
-  MaryImp.spec @era
-  describe "AlonzoImpSpec" . withEachEraVersion @era $ do
-    Bbody.spec
-    Utxo.spec
-    Utxos.spec
-    Utxow.spec
+spec era = do
+  Mary.spec era
+  describe "AlonzoEra Onwards" $ withImpInitEachEraVersion era $ do
+    BBODY.spec
+    UTXO.spec
+    UTXOS.spec
+    UTXOW.spec
 
 alonzoEraSpecificSpec ::
-  forall era.
   (AlonzoEraImp era, ShelleyEraTxCert era) =>
-  SpecWith (ImpInit (LedgerSpec era))
-alonzoEraSpecificSpec = do
-  describe "Alonzo era specific Imp spec" $
-    describe "Certificates without deposits" $ do
-      Utxow.alonzoEraSpecificSpec
-
-instance EraSpecificSpec AlonzoEra where
-  eraSpecificSpec =
-    ShelleyImp.shelleyEraSpecificSpec >> alonzoEraSpecificSpec
+  proxy era ->
+  Spec
+alonzoEraSpecificSpec era = withImpInitEachEraVersion era $ do
+  describe "AlonzoEra Specific" $ do
+    UTXOW.alonzoEraSpecificSpec