+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/UtxoSpec.hs
@@ -38,7 +38,7 @@ import Test.Cardano.Ledger.Imp.Common
 import Test.Cardano.Ledger.Plutus.Examples (alwaysSucceedsNoDatum)
 
 spec :: forall era. ConwayEraImp era => SpecWith (ImpInit (LedgerSpec era))
-spec = do
+spec = describe "UTXO" $ do
   describe "Certificates" $ do
     it "Reg/UnReg collect and refund correct amounts" $ do
       utxoStart <- getUTxO
@@ -157,7 +157,7 @@ conwayEraSpecificSpec ::
   , ShelleyEraTxCert era
   ) =>
   SpecWith (ImpInit (LedgerSpec era))
-conwayEraSpecificSpec = do
+conwayEraSpecificSpec = describe "UTXO" $ do
   describe "Certificates" $ do
     it "Reg/UnReg collect and refund correct amounts" $ do
       utxoStart <- getUTxO