+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/Imp.hs
@@ -1,7 +1,4 @@
-{-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeOperators #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
@@ -9,42 +6,31 @@
 module Test.Cardano.Ledger.Dijkstra.Imp where
 
 import Cardano.Ledger.Conway.Rules
-import Cardano.Ledger.Dijkstra (DijkstraEra)
 import Cardano.Ledger.Dijkstra.Core
 import Cardano.Ledger.Shelley.Rules
 import Test.Cardano.Ledger.Common
 import qualified Test.Cardano.Ledger.Conway.Imp as ConwayImp
-import qualified Test.Cardano.Ledger.Dijkstra.Imp.CertSpec as Cert
-import qualified Test.Cardano.Ledger.Dijkstra.Imp.CertsSpec as Certs
-import qualified Test.Cardano.Ledger.Dijkstra.Imp.LedgerSpec as Ledger
-import qualified Test.Cardano.Ledger.Dijkstra.Imp.UtxoSpec as Utxo
-import qualified Test.Cardano.Ledger.Dijkstra.Imp.UtxowSpec as Utxow
+import qualified Test.Cardano.Ledger.Dijkstra.Imp.CertSpec as CERT
+import qualified Test.Cardano.Ledger.Dijkstra.Imp.CertsSpec as CERTS
+import qualified Test.Cardano.Ledger.Dijkstra.Imp.LedgerSpec as LEDGER
+import qualified Test.Cardano.Ledger.Dijkstra.Imp.UtxoSpec as UTXO
+import qualified Test.Cardano.Ledger.Dijkstra.Imp.UtxowSpec as UTXOW
 import Test.Cardano.Ledger.Dijkstra.ImpTest
 
 spec ::
-  forall era.
   ( DijkstraEraImp era
-  , EraSpecificSpec era
   , Event (EraRule "EPOCH" era) ~ ConwayEpochEvent era
   , Event (EraRule "NEWEPOCH" era) ~ ConwayNewEpochEvent era
   , Event (EraRule "HARDFORK" era) ~ ConwayHardForkEvent era
   , Event (EraRule "RUPD" era) ~ RupdEvent
   ) =>
+  proxy era ->
   Spec
-spec = do
-  ConwayImp.spec @era
-  withEachEraVersion @era $ dijkstraEraGenericSpec @era
-
-dijkstraEraGenericSpec ::
-  forall era.
-  DijkstraEraImp era =>
-  SpecWith (ImpInit (LedgerSpec era))
-dijkstraEraGenericSpec = do
-  describe "LEDGER" Ledger.spec
-  describe "CERTS" $ do
-    Cert.spec
-    Certs.spec
-  describe "UTXOW" Utxow.spec
-  describe "UTXO" Utxo.spec
-
-instance EraSpecificSpec DijkstraEra
+spec era = do
+  ConwayImp.spec era
+  describe "DijkstraEra Onwards" $ withImpInitEachEraVersion era $ do
+    LEDGER.spec
+    CERT.spec
+    CERTS.spec
+    UTXOW.spec
+    UTXO.spec