+++ eras/shelley/impl/test/Main.hs
@@ -1,21 +1,25 @@
 {-# LANGUAGE TypeApplications #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
 
 module Main where
 
 import Cardano.Ledger.Shelley (ShelleyEra)
-import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Core.JSON (roundTripJsonEraSpec)
+import Test.Cardano.Ledger.Era
 import qualified Test.Cardano.Ledger.Shelley.Binary.CddlSpec as Cddl
 import qualified Test.Cardano.Ledger.Shelley.BinarySpec as Binary
 import qualified Test.Cardano.Ledger.Shelley.Imp as Imp
 import Test.Cardano.Ledger.Shelley.JSON (roundTripJsonShelleyEraSpec)
 
+instance EraSpec ShelleyEra where
+  eraImpSpec era = do
+    Imp.shelleyEraSpecificSpec era
+    Imp.spec era
+
 main :: IO ()
 main =
-  ledgerTestMain $
-    describe "Shelley" $ do
-      Binary.spec
-      Cddl.spec
-      Imp.spec @ShelleyEra
-      roundTripJsonEraSpec @ShelleyEra
-      roundTripJsonShelleyEraSpec @ShelleyEra
+  ledgerEraTestMain @ShelleyEra $ do
+    Binary.spec
+    Cddl.spec
+    roundTripJsonEraSpec @ShelleyEra
+    roundTripJsonShelleyEraSpec @ShelleyEra