+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Imp.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE ScopedTypeVariables #-}
@@ -9,49 +8,43 @@
 module Test.Cardano.Ledger.Shelley.Imp (spec, shelleyEraSpecificSpec) where
 
 import Cardano.Ledger.Core (EraRule)
-import Cardano.Ledger.Shelley (ShelleyEra)
 import Cardano.Ledger.Shelley.Rules (RupdEvent)
 import Cardano.Ledger.Shelley.State (ShelleyEraAccounts)
 import Control.State.Transition (Event)
 import Test.Cardano.Ledger.Imp.Common
-import qualified Test.Cardano.Ledger.Shelley.Imp.DelegSpec as Deleg
-import qualified Test.Cardano.Ledger.Shelley.Imp.EpochSpec as Epoch
-import qualified Test.Cardano.Ledger.Shelley.Imp.LedgerSpec as Ledger
-import qualified Test.Cardano.Ledger.Shelley.Imp.PoolSpec as Pool
-import qualified Test.Cardano.Ledger.Shelley.Imp.UtxoSpec as Utxo
-import qualified Test.Cardano.Ledger.Shelley.Imp.UtxowSpec as Utxow
+import qualified Test.Cardano.Ledger.Shelley.Imp.DelegSpec as DELEG
+import qualified Test.Cardano.Ledger.Shelley.Imp.EpochSpec as EPOCH
+import qualified Test.Cardano.Ledger.Shelley.Imp.LedgerSpec as LEDGER
+import qualified Test.Cardano.Ledger.Shelley.Imp.PoolSpec as POOL
+import qualified Test.Cardano.Ledger.Shelley.Imp.UtxoSpec as UTXO
+import qualified Test.Cardano.Ledger.Shelley.Imp.UtxowSpec as UTXOW
 import Test.Cardano.Ledger.Shelley.ImpTest
 import qualified Test.Cardano.Ledger.Shelley.UnitTests.InstantStakeTest as Instant
 
 spec ::
-  forall era.
+  forall proxy era.
   ( ShelleyEraImp era
-  , EraSpecificSpec era
   , Event (EraRule "RUPD" era) ~ RupdEvent
   ) =>
+  proxy era ->
   Spec
-spec = do
-  describe "Era specific tests" . withEachEraVersion @era $ eraSpecificSpec
-  describe "ShelleyImpSpec" $ withEachEraVersion @era $ do
-    describe "DELEG" Deleg.spec
-    Epoch.spec
-    Ledger.spec
-    Pool.spec
-    Utxow.spec
-    Utxo.spec
-  describe "ShelleyPureTests" $ do
+spec era = do
+  describe "ShelleyEra Onwards" $ withImpInitEachEraVersion era $ do
+    DELEG.spec
+    EPOCH.spec
+    LEDGER.spec
+    POOL.spec
+    UTXOW.spec
+    UTXO.spec
+  describe "ShelleyEraPureTests" $ do
     Instant.spec @era
 
 shelleyEraSpecificSpec ::
-  forall era.
   ( ShelleyEraImp era
   , ShelleyEraAccounts era
   ) =>
-  SpecWith (ImpInit (LedgerSpec era))
-shelleyEraSpecificSpec = do
-  describe "Shelley era specific Imp spec" $
-    describe "DELEG" $
-      Deleg.shelleyEraSpecificSpec
-
-instance EraSpecificSpec ShelleyEra where
-  eraSpecificSpec = shelleyEraSpecificSpec
+  proxy era ->
+  Spec
+shelleyEraSpecificSpec era = withImpInitEachEraVersion era $ do
+  describe "ShelleyEra Specific" $
+    DELEG.shelleyEraSpecificSpec