+++ eras/alonzo/impl/CHANGELOG.md
@@ -44,6 +44,8 @@
 
 ### `testlib`
 
+* Add `proxy era` argument to `alonzoEraSpecificSpec`
+* Make `Test.Cardano.Ledger.Alonzo.Imp.spec` accept `Proxy era`
 * Add `genDatumPresent`
 * Added `Proxy era` argument to `exUnitsRule`
 * `TranslationInstance` has a new field `tiPlutusPurpose`