+++ eras/conway/impl/CHANGELOG.md
@@ -46,6 +46,8 @@
 
 ### `testlib`
 
+* Export `conwayEraSpecificSpec`
+* Make `Test.Cardano.Ledger.Conway.Imp.spec` accept `Proxy era`
 * In `Test.Cardano.Ledger.Conway.Examples`:
   - Remove `mkConwayBasedExampleTx`, `exampleConwayBasedTxBody`
   - Add `exampleConwayBasedTx`, `exampleConwayBasedTopTx`